requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-phonenumber-field==7.0.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-extra-fields==3.4.1
idna==3.4
phonenumbers==8.13.4
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
django-user-agents==0.4.0
aiogram~=2.25.1
cachetools==3.1.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.13.1
oauth2client==4.1.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
rsa==4.0
uritemplate==3.0.0

pypdf2~=3.0.1
beautifulsoup4~=4.12.2
channels~=4.0.0